Add `.env_clear()` to all `Command` usages
This is mostly for completeness, since we do want to use the
current process's env vars, however we've already inherited
them in `main.rs` via `Env::from_current()`, and so the env
we're passing around is the exact env we want to use.

By adding `.env_clear()` we prevent subtle bugs if we ever
unset any of the inherited envs in the passed around `env`.