Don't skip the Check Changelog check for Python dependencies
Now that the Pip/Setuptools dependency versions are being tracked in
requirements files (in order to reduce the toil of keeping them up to date,
since Dependabot can open PRs to update them), this means that it's
not correct to always skip all Dependabot PRs, since the Python subset
will still want a changelog entry.

GUS-W-13619167.