Always add the Pip dependencies layer `bin` directory to PATH
Previously the `bin` directory would only be added to `PATH` if it
existed (due to it being set by `lifecycle`), which is only the case
if one of the installed dependencies has an entry point script, which
causes Pip to create the `bin` directory and place a wrapper script
inside it for that package's CLI command.

This meant:
1. We had to suppress Pip's warning during install (since Pip isn't
  to know that the `PATH` will be correctly set later)
2. If an app has no dependencies with an entry-point (so the app
  image doesn't have a `bin` directory), then at run-time tries to
  pip install a new package that does have an entry point (eg when
  debugging), then that new script won't be on `PATH`.

As such, we now add the `bin` directory to PATH explicitly, instead
of relying on `lifecycle` automatic addition.