Update tests for Python 3.9 on Ubuntu 24.04
Previously Python 3.9 was only supported on the Heroku-22
stack (Ubuntu 22.04) and older, since it reaches EOL in Oct 2025.

However, support for Python 3.9 on Heroku-24 (Ubuntu 24.04) was
just added, by the building/uploading of the binaries in:
https://github.com/heroku/heroku-buildpack-python/pull/1656

As such, the tests need updating to reflect that CNB builds will no
longer fail with an unavailable Python error when using Python 3.9
on `heroku/builder:24`.

GUS-W-16917997.