Remove support for `runtime.txt`
The Heroku-proprietary `runtime.txt` file was previously
deprecated in both the classic buildpack and CNB, in favour
of the more widely supported `.python-version` file:
https://devcenter.heroku.com/changelog-items/3005
https://devcenter.heroku.com/changelog-items/3141

We will be keeping support for `runtime.txt` in the classic
Python buildpack for a while longer. However, since the CNB
is new (and not used by any/many Heroku apps yet since it only
GAed this week), we're dropping support for `runtime.txt` here
sooner, to encourage migration. (Particularly since uv support
will require apps to be using `.python-version`)

Closes #275.
GUS-W-16878263.