Improve error messages for invalid version in `.python-version` (#353)
This ports some of the recent error message improvements made in the
classic Python buildpack to the CNB, plus some additional tweaks based
on Honeycomb metrics for classic (which showed quite a bit of attempted
usage of `.x` and `.*` style wildcards, which aren't valid syntax for
the `.python-version` file).

GUS-W-18346802.