Improve the error messages for file I/O errors (#355)
This improves the error messages shown for file related I/O errors,
by:
- ensuring that the path is always included
- adding additional suggestions to the error based on the context
  (which is now known, since specialised types are used for each of
  the common scenarios like checking if a file exists or reading a
  file to a string)

We're able to make the errors richer than those provided by `fs-err`,
since we know the context in which the errors have occurred.

For example, in the error message for `ReadOptionalFileError`s we know
not to say to check for a missing file or that the file is a directory,
since `read_optional_file()` explicitly excludes those cases by design.
If we used `io::Error` (or `fs-err`'s equivalent) we would have to rely
on the user not using the wrong logging function for the wrong I/O error
case, rather than relying on the type system to do that for us.

These changes leave `log_io_error()` as being used only for the
`Command` I/O cases. A later PR will switch adjust those usages too
(but this PR is already large enough).

Closes #270.
GUS-W-12650236.