Improve the error messages for command I/O errors
Similar to #355 but for I/O errors from spawning a `Command`
instead of from file operations.

Such Command I/O errors are only likely to happen in the case
of a bug, so the error message now says as much. The error
also now generates the program name from the command,
so avoid it getting out of sync. The command's args are
intentionally omitted, since they aren't relevant for errors
that occur when spawning a command. (ie: For command not
found, it's the program that's not found, not the args.)

GUS-W-12650236.