Update test assertions after lifecycle output change
`lifecycle` now emits all buildpack output to stdout in order
to fix this stdout/stderr output interleaving bug:
https://github.com/buildpacks/pack/issues/2330
https://github.com/buildpacks/lifecycle/pull/1525

As such, the `libcnb-test` integration test assertions need
updating to use stdout in a number of places.

This also allows the test assertions to correctly assert the
placement of warning in-between other non-warning
content. (And in the process has revealed a couple of places
where whitespace could be tweaked, which will be handled
separately as part of the future build output style change work).

GUS-W-19337364.