Deprecate support for Python 3.10 (#467)
Since Python 3.10 reaches upstream EOL in October 2026:
https://devguide.python.org/versions/#supported-versions

The Python version support policy is that supported versions
follows the upstream EOL lifecycle:
https://devcenter.heroku.com/articles/python-support#python-version-support-policy

To raise awareness of this, a deprecation warning is now shown,
similar to that for Python 3.9.

In addition the Python 3.9 warning has been updated to use
the past tense to refer to the upstream EOL date, since it's
now in the past.

GUS-W-17595558.