Explicitly disable pip version checks when bootstrapping pip (#470)
These are currently implicitly disabled due to a pip implementation
detail, where it doesn't check for updates unless it finds that it is
installed in the current environment.

However, it's clearer if we disable them explicitly (and it also saves
pip from having to query the list of currently installed packages).

See:
https://pip.pypa.io/en/stable/cli/pip/#cmdoption-disable-pip-version-check

GUS-W-20289843.